fix: add log a1

+++ bpf/process/types/basic.h
@@ -520,7 +520,7 @@ __copy_char_buf(long off, unsigned long arg, unsigned long bytes,
 	int err;
 
 	/* Bound bytes <4095 to ensure bytes does not read past end of buffer */
-	size_t tmp = 1024;
+	unsigned tmp = 1024 ;
 	rd_bytes = bytes;
 	rd_bytes = tmp < rd_bytes ? tmp : rd_bytes;
 	err = probe_read(&s[2], rd_bytes, (char *)arg);
